First version that raises no warnings
This is the first version after re-naming the functions and the package to stabest that raises no warnings from R CMD CHECK. However, https://win-builder.r-project.org/ still raises a warning because there are "No examples, no tests, no vignettes". Also, a number of warnings occur in compiling the C code - mostly because signed and unsigned integers are compared in some assert() statements. This should be fixed in the future.