pass latent valuations as initial parameter
Added an option to pass the latent valuation of students over schools (Vc) and vice versa (Vs) as an initial parameter. To this end, one can now include variables Vc_ and Vs_ in the dataset that is provided to stabest. If these variables are present, than their values will be used instead of the automatic initializations.